+++ backend/canisters/group/impl/src/updates/c2c_install_bot.rs
@@ -22,6 +22,7 @@ fn c2c_install_bot_impl(args: Args, state: &mut RuntimeState) -> OCResult {
     }
 
     let installed_by = member.user_id();
+    let now = state.env.now();
 
     if !state.data.install_bot(
         installed_by,
@@ -29,7 +30,7 @@ fn c2c_install_bot_impl(args: Args, state: &mut RuntimeState) -> OCResult {
         args.granted_permissions.clone(),
         args.granted_autonomous_permissions.clone(),
         args.default_subscriptions,
-        state.env.now(),
+        now,
     ) {
         return Err(OCErrorCode::AlreadyAdded.into());
     }
@@ -43,6 +44,7 @@ fn c2c_install_bot_impl(args: Args, state: &mut RuntimeState) -> OCResult {
             granted_autonomous_permissions: args.granted_autonomous_permissions.unwrap_or_default(),
         })),
         recipients: vec![args.bot_id],
+        timestamp: now,
     }));
 
     handle_activity_notification(state);